Added Touch Controls. Updated Buttons. Fixed Music. (#25)
I added touch controls for left, right, and jump. I updated the skip cinematic button (gave it a theme). I updated the map UI to have settings instead of "menu" and when you click it, it pauses the game. You are able to restart level and quit to menu from inside the settings page in game.